beamfoundry/requirements.txt
Flask==3.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
matplotlib==3.10.8
psycopg2-binary==2.9.9
pydantic==2.12.5
python-dotenv==1.0.1
pyserial==3.5
pytest==8.2.1
pytz==2025.2

# Type checking
mypy

# Formatting
black

# Testing utilities
freezegun

# Linting
ruff==0.6.9
